small refactor and fixes
