+++ App.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import {
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+const App = () => (
+  <View style={styles.container}>
+    <Text>Open up App.js to start working on your app!</Text>
+    <Text>Changes you make will automatically reload.</Text>
+    <Text>Shake your phone to open the developer menu.</Text>
+  </View>
+);
+
+export default App;
