+++ App.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import App from './App';
-
 import renderer from 'react-test-renderer';
+
+import App from './App';
 
 it('renders without crashing', () => {
   const rendered = renderer.create(<App />).toJSON();
