+++ package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "expo": "^21.0.0",
     "react": "16.0.0-alpha.12",
-    "react-native": "^0.48.4"
+    "react-native": "^0.49.3"
   },
   "devDependencies": {
     "eslint": "^4.9.0",
